Air quality completely displaying, made info alert dialog

+++ app/src/main/java/com/personal/weatherapp/data/remote/WeatherDataDto.kt
@@ -13,7 +13,7 @@ data class WeatherDataDto(
     @field:Json(name = "windspeed_10m")
     val windSpeeds: List<Double>,
     @field:Json(name = "relativehumidity_2m")
-    val humidities: List<Double>,
+    val humidities: List<Int>,
     @field:Json(name = "is_day")
     val isDay: List<Int>
 )
